rename reducers path to (reducers)
