+++ src/components/paginator.js
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from 'react';
 import '../index.css';
 
 const Paginator = ({ totalMails, onPageChange, pagination, mailsPerPage }) => {
-  const [counter, setCounter] = useState(1);
 
-  const value = mailsPerPage * counter + 1 ;
+  const [counter, setCounter] = useState(1);
+  const value = mailsPerPage * counter + 1;
 
   useEffect(() => {
-    onPageChange(value -mailsPerPage, value);
+    onPageChange(value - mailsPerPage, value);
   }, [counter]);
 
   const onButtonClick = (type) => {
@@ -16,7 +16,7 @@ const Paginator = ({ totalMails, onPageChange, pagination, mailsPerPage }) => {
         setCounter(counter - 1);
       }
     } else if (type === 'next') {
-      if (counter + 1  < Math.ceil(totalMails/mailsPerPage)) {
+      if (counter + 1 < Math.ceil(totalMails / mailsPerPage)) {
         setCounter(counter + 1);
       }
     }
